backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
openai
google-generativeai
python-docx
openpyxl
python-multipart
pymupdf
